Harden server actions and lesson conversations
- Restrict admin content mutations and user listing
- Rate-limit and authenticate lesson AI calls
- Move webhook user sync and prompt helpers out of public client/action paths
